L4G-8 Add testcase for Login API calls [RDZ, FC]
