--- Client/app/src/main/java/com/Client/Table/data/model/User.kt
@@ -1,4 +1,0 @@
-package com.Client.Table.data.model
-
-data class User(val username: String, val email: String, val password: String)
-data class RegistrationResponse(val id: Int)
