+++ Client/app/src/main/java/com/Client/Table/ui/registration/RegistrationActivity.kt
@@ -1,5 +1,6 @@
 package com.Client.Table.ui.registration
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Patterns
@@ -12,6 +13,7 @@ import com.Client.Table.R
 import androidx.lifecycle.viewModelScope
 import com.Client.Table.data.model.User
 import com.Client.Table.network.BackendApi
+import com.Client.Table.ui.login.LoginActivity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -112,6 +114,8 @@ class RegistrationActivity : AppCompatActivity() {
 
             if (success) {
                 Toast.makeText(this, "Registration Successful", Toast.LENGTH_SHORT).show()
+                val intent = Intent(this, LoginActivity::class.java).apply{}
+                startActivity(intent)
             } else {
                 Toast.makeText(this, "Registration Unsuccessful", Toast.LENGTH_SHORT).show()
             }
